+++ cmpc/script_tester.py
@@ -1,45 +1,32 @@
 import asyncio
-import twitchio
-import twitchio.dataclasses as dataclasses
+
 
 class MockUser:
     def __init__(self, name, id: int = 0):
-        self._name = name
-        self._id = id
-    
-    @property
-    def name(self) -> str:
-        return self._name
+        self.name = name
+        self.id = id
 
-    @property
-    def id(self) -> int:
-        return self._id
 
 class MockMessage:
-    def __init__(self, author: MockUser, messageContent: str):
-        self._content = messageContent
-        self._author = author
-    
-    @property
-    def content(self) -> str:
-        return self._content
+    def __init__(self, author: MockUser, content: str):
+        self.content = content
+        self.author = author
 
-    @property
-    def author(self) -> MockUser:
-        return self._author
 
 class ScriptTester(object):
-    def __init__ (self, messageCallback, originalSelf):
-        self.messageCallback = messageCallback
+    def __init__(self, messageCallback, originalSelf):
+        self.message_callback = messageCallback
         self._user = None
-        self._originalSelf = originalSelf
+        self._original_self = originalSelf
+        self._last_message = None
 
-    async def inputLoop(self):
+    async def input_loop(self):
         while True:
-            if self._user == None:
+            if self._user is None:
                 user = input('[Offline Mode] What user would you like to mock?')
-                userid = input('[Offline Mode] What user ID would you like to use? (If you dont know what this is, use 0)')
-                if userid != "":
+                userid = input('[Offline Mode] What user ID would you like to use? '
+                               "(If you don't know what this is, use 0)")
+                if userid:
                     self._author = MockUser(user, userid)
                 else:
                     self._author = MockUser(user)
@@ -50,14 +37,13 @@ class ScriptTester(object):
             elif message == '//exit':
                 print('[Offline Mode] Use Ctrl-C to exit (Command-C on MacOS)')
             else:
-                self._lastMessage = MockMessage(self._author, message)
-                await self.messageCallback(self=self._originalSelf, message=self._lastMessage)
+                self._last_message = MockMessage(self._author, message)
+                await self.message_callback(self=self._original_self, message=self._last_message)
 
-
-    def startTester(self):
+    def run(self):
         """Starts the input loop which makes a mock twitchio message, then sends it to the messageCallback."""
         loop = asyncio.get_event_loop()
-        loop.create_task(self.inputLoop())
+        loop.create_task(self.input_loop())
 
         try:
             print('[Offline Mode] Starting Input Loop')
@@ -68,6 +54,4 @@ class ScriptTester(object):
         print('[Offline Mode] Stopping Asyncio Loop')
         loop.stop()
         loop.close()
-        
-